requirements.txt
fastapi
uvicorn[standard]
openai
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
pydantic
ffmpeg-python
requests
